create libk
