+++ kernel/Makefile
@@ -7,14 +7,15 @@ LINKER_PATH = src/ld/linker.ld
 
 MKDIR_P ?= mkdir -p
 
-CFLAGS = --target=i386-pc-none-elf -ffreestanding -O2 -std=c99 -Wall -Wextra
+INCLUDE_FLAGS = -I./include -I../libc/include -I../libk/include
+CFLAGS = --target=i386-pc-none-elf -ffreestanding -O2 -std=c99 -Wall -Wextra $(INCLUDE_FLAGS)
 ASFLAGS = -felf32
-LDFLAGS = -nostdlib
+LDFLAGS = -nostdlib -L../libc/build/ -L../libk/build -lc -lk
 
 all: build/os.bin
 
 build/os.bin: ./build/kernel.c.o ./build/boot.asm.o
-	clang -T $(LINKER_PATH) $(CFLAGS) $^ -o $@ $(LDFLAGS) -L../libc/build/ -lc
+	clang -T $(LINKER_PATH) $(CFLAGS) $^ -o $@ $(LDFLAGS)
 
 $(BUILD_DIR)/%.asm.o: $(SRC_ASM_DIR)/%.asm
 	$(MKDIR_P) $(dir $@)
@@ -22,7 +23,7 @@ $(BUILD_DIR)/%.asm.o: $(SRC_ASM_DIR)/%.asm
 
 $(BUILD_DIR)/%.c.o: $(SRC_C_DIR)/%.c
 	$(MKDIR_P) $(dir $@)
-	clang $(CFLAGS) -I../libc/include -c $< -o $@
+	clang $(CFLAGS) -c $< -o $@
 
 clean:
 	$(RM) -r $(BUILD_DIR)
