+++ libc/include/stdio.h
@@ -1,0 +1,9 @@
+
+#ifndef OS_STDIO_H
+#define OS_STDIO_H
+
+int printf(const char* __restrict, ...);
+int putchar(int);
+int puts(const char*);
+
+#endif //OS_STDIO_H
