+++ libc/src/c/stdio/printf.c
@@ -1,0 +1,4 @@
+
+int printf(const char* __restrict s, ...) {
+    return 0;
+}
