+++ libc/src/c/stdio/putc.c
@@ -1,0 +1,6 @@
+
+#include <terminal.h>
+
+int putchar(char c) {
+    return terminal_putchar(c);
+}
