+++ libc/src/c/stdio/puts.c
@@ -1,0 +1,6 @@
+
+#include <terminal.h>
+
+int puts(const char* s) {
+    return terminal_puts(s);
+}
