+++ libk/src/c/terminal/write.c
@@ -1,0 +1,48 @@
+
+#include "terminal.h"
+#include "color.h"
+#include <stddef.h>
+
+size_t k_strlen(const char* str) {
+    size_t len = 0;
+    while (str[len])
+        len++;
+    return len;
+}
+
+void terminal_setcolor(uint8_t color) {
+    terminal_color = color;
+}
+
+void terminal_putentryat(char c, uint8_t color, size_t x, size_t y) {
+    const size_t index = y * VGA_WIDTH + x;
+    terminal_buffer[index] = vga_entry(c, color);
+}
+
+int terminal_putchar(char c) {
+    if(c == '\n') {
+        terminal_column = 0;
+        terminal_row++;
+    } else {
+        terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
+        if (++terminal_column == VGA_WIDTH) {
+            terminal_column = 0;
+            if (++terminal_row == VGA_HEIGHT)
+                terminal_row = 0;
+        }
+    }
+    return c;
+}
+
+void terminal_write(const char* data, size_t size) {
+    for (size_t i = 0; i < size; i++)
+        terminal_putchar(data[i]);
+}
+
+int terminal_puts(const char* data) {
+    int len = k_strlen(data);
+    terminal_write(data, len);
+
+    return len;
+}
+
